requirements.txt
mysql-connector==2.2.9
python-decouple==3.5
xmltodict==0.12.0
